[#6] refactor: molecules component import 경로 전환
