+++ frontend/src/components/molecules/Header/Header.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import Header from './Header';
-import { Page } from '../../../schemes';
+import { Page } from '@/schemes';
+import Header from '.';
 
 const desc = {
   component: Header,
