+++ frontend/src/components/molecules/index.ts
@@ -1,4 +1,5 @@
-export * as Header from './Header';
-export * as Editor from './Editor';
-export * as Menu from './Menu';
-export * as BlockComponent from './BlockComponent';
+export { default as Header } from './Header';
+export { default as Editor } from './Editor';
+export { default as Menu } from './Menu';
+export { default as BlockComponent } from './BlockComponent';
+export { default as BlockHandler } from './BlockHandler';
